python package requirements.txt
absl-py==0.2.0
alabaster==0.7.10
anaconda-client==1.6.3
anaconda-navigator==1.6.4
anaconda-project==0.6.0
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.22.0
astor==0.6.2
astroid==1.5.3
astropy==1.3.2
awscli==1.15.31
Babel==2.5.0
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0rc1
beautifulsoup4==4.6.0
bitarray==0.8.1
blaze==0.10.1
bleach==1.5.0
bokeh==0.12.5
boto==2.46.1
boto3==1.7.31
botocore==1.10.31
Bottleneck==1.2.1
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
cloudpickle==0.2.2
clyent==1.2.2
colorama==0.3.7
contextlib2==0.5.5
cryptography==1.8.1
cvxpy==1.0.3
cycler==0.10.0
Cython==0.26.1
cytoolz==0.8.2
d360util===v4.0.1mp-125-g4c45b73-dirty
dask==0.14.3
datashape==0.5.4
decorator==4.1.2
dill==0.2.7.1
distributed==1.16.3
Django==2.0.3
docutils==0.14
easydict==1.7
ecos==2.0.5
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
fbprophet==0.1.1
Flask==0.12.2
Flask-Cors==3.0.2
future==0.16.0
fuzzywuzzy==0.15.1
gast==0.2.0
geopy==1.11.0
gevent==1.2.1
graphviz==0.8
greenlet==0.4.12
grpcio==1.11.0
h5py==2.7.0
haversine==0.4.5
HeapDict==1.0.0
html5lib==0.9999999
idna==2.6
image==1.5.19
image-slicer==0.1.1
imagesize==0.7.1
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.1.0
jupyter-core==4.3.0
Keras==2.1.6
keras-resnet==0.1.0
keras-retinanet==0.0.1
lazy-object-proxy==1.3.1
llvmlite==0.18.0
locket==0.2.0
lxml==3.7.3
Markdown==2.6.8
MarkupSafe==1.0
matplotlib==2.0.2
mccabe==0.6.1
mistune==0.7.4
mock==2.0.0
mpmath==0.19
msgpack-python==0.4.8
multipledispatch==0.4.9
multiprocess==0.70.5
mxnet==0.11.0
navigator-updater==0.1.0
nbconvert==5.3.0
nbformat==4.4.0
networkx==1.11
nltk==3.3
nose==1.3.7
notebook==5.0.0
numba==0.33.0
numexpr==2.6.2
numpy==1.14.0
numpydoc==0.7.0
odo==0.5.0
olefile==0.44
opencv-python==3.4.1.15
openpyxl==2.4.7
osqp==0.3.0
packaging==16.8
pandas==0.20.3
pandocfilters==1.4.2
partd==0.3.8
pathlib2==2.2.1
patsy==0.4.1
pbr==3.1.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
plotly==2.7.0
ply==3.10
prompt-toolkit==1.0.14
prophet==0.1.1
protobuf==3.5.2.post1
psutil==5.3.0
psycopg2==2.7.3.1
ptyprocess==0.5.1
PuLP==1.6.8
py==1.4.33
pyasn1==0.4.2
pycodestyle==2.3.1
pycosat==0.6.2
pycparser==2.17
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.7.2
pyodbc==4.0.16
pyOpenSSL==17.0.0
pyparsing==2.1.4
pyshp==1.2.11
pystan==2.16.0.0
pytest==3.0.7
python-dateutil==2.6.1
python-Levenshtein==0.12.0
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
rake-nltk==1.0.1
requests==2.18.4
rope==0.10.7
rope-py3k==0.9.4.post1
rsa==3.4.2
s3fs==0.1.2
s3transfer==0.1.12
scikit-image==0.13.0
scikit-learn==0.18.1
scipy==0.19.0
scs==2.0.2
seaborn==0.7.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
SQLAlchemy==1.1.9
statsmodels==0.8.0
sympy==1.0
tables==3.3.0
tblib==1.3.2
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
terminado==0.6
testpath==0.3.1
Theano==0.9.0
toolz==0.8.2
tornado==4.5.1
tqdm==4.19.4
traitlets==4.3.2
unicodecsv==0.14.1
urllib3==1.22
uszipcode==0.1.3
versioneer==0.18
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==2.0.0
wrapt==1.10.11
xlrd==1.0.0
XlsxWriter==0.9.6
xlwings==0.10.4
xlwt==1.2.0
zict==0.1.2
zipcode==2.0.0
